requirements.txt
asgiref==3.2.3
Django==3.0.2
ffmpeg==1.4
ffmpy==0.2.2
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-magic==0.4.15
pytz==2019.3
sqlparse==0.3.0
